Updated Makefile
Updates Makefile for use on updated Iowa State servers (a new version of CPLEX was installed). This is probably an arbitrary option since the CPLEX path is going to vary by installation, but since Iowa State is currently the only place that this is used, it may be good for Netplan to compile right out of the box.